src: add support for multiple sensors (#11)
This new changes make future sensor integration to Application server
become easier to maintain and upgrade. The Application server now need
to provide the FPORT when it want to fetch the data, a FPORT map to
an Application, for example port 1 is temperature and humidity sensor.

Future sensor integration should be done in a seperate tab in the Navbar.
This new change add support for temperature and humidity.

Check the src/App.js to see how to implement different tabs. Developers
are expected to integrate logic in the useFetchSensorData hook so that
it retrieve based on FPORT.